Finished the project
Made a new component so the user could switch between validation type and dark mode of the app.

Added some styles and refinement for the formC component, added 2 instances of the newly created switch component to the App View, also added all of the styling for all the components!

Made a file for the unit tests and made the basic tests required to invalidate the form.

Modified readme file